+++ lib/src/windows/bluetooth_characteristic_windows.dart
@@ -34,13 +34,11 @@ class BluetoothCharacteristicWindows extends BluetoothCharacteristic {
             properties: BmCharacteristicProperties(
               broadcast: propertiesWinBle.broadcast ?? false,
               read: propertiesWinBle.read ?? false,
-              writeWithoutResponse:
-                  propertiesWinBle.writeWithoutResponse ?? false,
+              writeWithoutResponse: propertiesWinBle.writeWithoutResponse ?? false,
               write: propertiesWinBle.write ?? false,
               notify: propertiesWinBle.notify ?? false,
               indicate: propertiesWinBle.indicate ?? false,
-              authenticatedSignedWrites:
-                  propertiesWinBle.authenticatedSignedWrites ?? false,
+              authenticatedSignedWrites: propertiesWinBle.authenticatedSignedWrites ?? false,
               // TODO: implementation missing
               extendedProperties: false,
               // TODO: implementation missing
@@ -59,8 +57,7 @@ class BluetoothCharacteristicWindows extends BluetoothCharacteristic {
   ///   - anytime `read()` is called
   ///   - anytime `write()` is called
   ///   - anytime a notification arrives (if subscribed)
-  List<int> get lastValue =>
-      FlutterBluePlusWindows._lastChrs[remoteId]?[_key] ?? [];
+  List<int> get lastValue => FlutterBluePlusWindows._lastChrs[remoteId]?[_key] ?? [];
 
   /// this stream emits values:
   ///   - anytime `read()` is called
@@ -74,14 +71,9 @@ class BluetoothCharacteristicWindows extends BluetoothCharacteristic {
             serviceId: serviceUuid.str128,
             characteristicId: characteristicUuid.str128,
           ),
-          FlutterBluePlusWindows._charReadWriteStream
-              .where((e) => e.$1 == _key)
-              .map((e) => e.$2)
+          FlutterBluePlusWindows._charReadWriteStream.where((e) => e.$1 == _key).map((e) => e.$2)
         ],
-      )
-          .map((p) => <int>[...p])
-          .newStreamWithInitialValue(lastValue)
-          .asBroadcastStream();
+      ).map((p) => <int>[...p]).newStreamWithInitialValue(lastValue).asBroadcastStream();
 
   /// this stream emits values:
   ///   - anytime `read()` is called
@@ -93,15 +85,12 @@ class BluetoothCharacteristicWindows extends BluetoothCharacteristic {
             serviceId: serviceUuid.str128,
             characteristicId: characteristicUuid.str128,
           ),
-          FlutterBluePlusWindows._charReadWriteStream
-              .where((e) => e.$1 == _key)
-              .map((e) => e.$2)
+          FlutterBluePlusWindows._charReadWriteStream.where((e) => e.$1 == _key).map((e) => e.$2)
         ],
       ).map((p) => <int>[...p]).asBroadcastStream();
 
   // TODO: need to verify
-  bool get isNotifying =>
-      FlutterBluePlusWindows._isNotifying[remoteId]?[_key] ?? false;
+  bool get isNotifying => FlutterBluePlusWindows._isNotifying[remoteId]?[_key] ?? false;
 
   Future<List<int>> read({int timeout = 15}) async {
     final value = await WinBle.read(
@@ -110,23 +99,22 @@ class BluetoothCharacteristicWindows extends BluetoothCharacteristic {
       characteristicId: characteristicUuid.str128,
     );
     FlutterBluePlusWindows._charReadWriteStreamController.add((_key, value));
+    FlutterBluePlusWindows._lastChrs[remoteId] ??= {};
     FlutterBluePlusWindows._lastChrs[remoteId]?[_key] = value;
     return value;
   }
 
   Future<void> write(List<int> value,
-      {bool allowLongWrite = false,
-      bool withoutResponse = false,
-      int timeout = 15}) async {
+      {bool allowLongWrite = false, bool withoutResponse = false, int timeout = 15}) async {
     await WinBle.write(
       address: _address,
       service: serviceUuid.str128,
       characteristic: characteristicUuid.str128,
       data: Uint8List.fromList(value),
-      writeWithResponse:
-          !withoutResponse, // propertiesWinBle.writeWithoutResponse ?? false,
+      writeWithResponse: !withoutResponse, // propertiesWinBle.writeWithoutResponse ?? false,
     );
     FlutterBluePlusWindows._charReadWriteStreamController.add((_key, value));
+    FlutterBluePlusWindows._lastChrs[remoteId] ??= {};
     FlutterBluePlusWindows._lastChrs[remoteId]?[_key] = value;
   }
 
@@ -157,6 +145,7 @@ class BluetoothCharacteristicWindows extends BluetoothCharacteristic {
           characteristicId: characteristicUuid.str128,
         );
       }
+      FlutterBluePlusWindows._isNotifying[remoteId] ??= {};
       FlutterBluePlusWindows._isNotifying[remoteId]?[_key] = notify;
     } catch (e) {
       log(e.toString());